Update patch set 23
Patch Set 23:

> Patch Set 23:
> 
> Swagger says this because there is no defined semantic for a body in a DELETE statement: https://tools.ietf.org/html/rfc7231#page-29
> 
> " A payload within a DELETE request message has no defined semantics;
>    sending a payload body on a DELETE request might cause some existing
>    implementations to reject the request."
> 
> probably seen as not-great HTTP design practice.
> 
> (I'm only adding this comment because I was added to this PR)

The 3 errors that the swagger editor detects are about the requestBody for DELETE.  That error is arguable in general - here is no standard and the swagger decided to make that as error in 3.x.  However, the editor does not prevent it from trying and sends the requestBody on DELETE just fine.  

Essentially, LUM needs to know the userId who revokes the record.  
Here are the options.  
1. keep the requestBody on DELETE - that is there 
2. move userId to query as param like ?userId=xxx.  -- becomes not case sensitive 
3. move userId to header as "userId"="xxx"

#2 and #3 break the pattern with other messages, but that would be openapi to blame.

Patch-set: 23