Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Looking at the build results, I see a lot of warnings about allOf and multiple inline schemas.  Is that OK?

what you see is about the java client code generation form openapi spec.  That piece is way behind the openapi standard and not capable of the points you mention.  However, since those are just warnings, that seems to be ok.  

LUM-server does not use java and does not use any code generation from openapi spec.  So, that does not affect the lum-server that this repo is about

Patch-set: 1
Label: Code-Review=0