Update patch set 2
Patch Set 2:

Hi Alex, Parag,

I have made fixes to the lum java client. There was an issue where the software creators was not able to be sent due to issues that are warnings in the logs with allOf missing types. When there is a property assigned to allOf type the code gen doesn't do a good job. We have to add a workaround type and then create a schema that uses the allOf code. This work around satisfies the swid tag creation. 

Justin

Patch-set: 2