Update patch set 2
Patch Set 2:

A node project, the LUM server, is not a Java project, so for me it is not a maven project.

Often we segregate builds when a repo has multiple things.  The server can be built by a job separately from the client.  If the builds are segregated, then the server can be released separately from the client, which can be VERY useful when you need to fix bugs but not make a breaking API change.  If you have a single build process then you have to release them in lockstep, which has tripped up several projects, they often forget to bump the versions on the components that did NOT change.

This is a long way of saying, I don't find it helpful to just run "mvn" at the root and have everything happen.  Might be convenient for a user, but not needed for Jenkins.

Patch-set: 2