Update patch set 7
Patch Set 7:

I think the javadoc publisher assumes the files were generated starting in the repo root directory; it cannot handle a project that's nested in a subdirectory as you have done here. So, yet another brutal assumption by LF.  we might finesse the problem by adding a tiny pom file to the parent directory with only a single submodule as its entry,  then the maven task could start in the root, and this publish might work.

Patch-set: 7