Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> parameters in a path should identify a specific object or collection.
> /dragons - the collection of all dragons
> /dragons/smaug - the dragon in the Hobbit
> /dragons/puff - the magic dragon
> 
> parameters in a query should control the results of the query
> /dragons?count=5 - don't list more than 5 dragons
> /dragons?ability=magic - only list dragons that can do magic
> 
> In Acumos Federation, for example:
> /catalogs - list all catalogs
> /solutions?catalogId=xxx - all solutions listed in catalog xxx
> /solutions/xxx - solution xxx
> /revision/xxx/documents?catalogId=yyy - revision xxx's documents that appear in catalog yyy
> 
> hope that helps.

I agree with Andrew except that there is a problem with passing a  iri/uri in a path variable. The problem is with how nginx reverse proxy will decode the uri to normalize it causing the service to not receive the encoded path. IMO..That is the only case that requires a change to use a query parameter.

Patch-set: 1
Label: Code-Review=0