Update patch set 6
Patch Set 6:

I submitted patch to downgrade the javadoc plugin and we are working again. Thanks Chris for the pointer to the javadoc plugin version for simple fix. 

Next steps for Demeter License Manager Client Library (LMCL) contributor is to take these changes and test them with the LMCL. In Clio I would setup a build locally and start the lum in my local docker and run tests against the repo. Michelle can setup meeting if we want to do it together. If this is merged in before this step the version may require some adjustments and Alex would need to provide additional updates. We like to have customer (LMCL) of LUM to give +1 here in Clio.

Patch-set: 6
Label: Code-Review=0