Update patch set 5
Patch Set 5: Code-Review-1

I don't understand this use of the prefix "pre."  Does it mean that these are things the user is not currently entitled to, but will be at some future time, that is, where there is a license, but its start date is in the future?  I can understand unentitled - you don't have a right to something, and entitled - you do have the right, but "preentitled" would mean before you have a right to do something, which would say you don't currently have the right.  This seems very confusing.

Patch-set: 5
Label: Code-Review=-1