Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> I don't understand this use of the prefix "pre."  Does it mean that these are things the user is not currently entitled to, but will be at some future time, that is, where there is a license, but its start date is in the future?  I can understand unentitled - you don't have a right to something, and entitled - you do have the right, but "preentitled" would mean before you have a right to do something, which would say you don't currently have the right.  This seems very confusing.

You have a good point !  I struggle with finding the right word.  

I need a verb that differs from entitled because "entitled" means yes entitled and the usage counter has been incremented.  This new "preentitled" API is different in that it kinda entitles, but does not increment the usage counters since it is not clear on which of many swidTags are going to be picked for the actual asset-usage.  In other words, it says yep - here is the list of kinda entitled software items, but to get the real entitlement on the asset-usage you (client of LUM) has to ask again on /asset-usage and it might be that you will get the entitlement or might be not if the usage counter exceeded already by that time or permission expired or etc.  For AcuCompose that works with the items in catalog that are not assets per se since no copy of the software is created when the composition happens, there is no usage counters and calling this API is good enough.  However, if any other client of LUM will try using this new API, I need to emphasize to them that this is not a real entitlement.  

Any better word suggestions than preentitled ?

Patch-set: 5