Update patch set 5
Patch Set 5:

++

Please help with wording for the new API - see the previous comment from Andrew and my answer.  

Need a synonym for "entitled" that would mean that it is a candidate for "entitled" but not really "entitled" yet because LUM does not count that as the asset-usage but only gives a hint that is a list of software items to pick from.  

Currently, I have 'preentitled" and that seems to be not the best choice.

thank you

Patch-set: 5