Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Alex, you need to fix the problem of having the version number appear in so many places.  I'm going to approve this one, but this is the last time.

Andrew - thank you for +1 and the comment.

I think that to "fix" the version-everywhere would require the change on CICD side first of all and probably introduction of the macro-replacement in so many files.  I did not see any good example of that so far.  It is annoying and to the developer probably more than to reviewer.  

Do you have any suggestions on how to do that ?

Patch-set: 1