Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Andrew,
> 
> I see that *-SNAPSHOT tags were pushed into nexus
> 
> docker push nexus3.acumos.org:10004/acumos/lum-server:1.3.2-SNAPSHOT
> docker push nexus3.acumos.org:10004/acumos/lum-db:1.3.2-SNAPSHOT
> 
> in
> https://jenkins.acumos.org/job/license-usage-manager-server-docker-merge-master/19/console
> https://jenkins.acumos.org/job/license-usage-manager-db-docker-merge-master/19/console

True.  So what it did is consistent with what happens when you build a Java based image.

But we don't have an equivalent to the job that runs on stage-release for Java based images that re-runs the build converting all references to snapshot versions to references to non-snapshot versions so that the resulting image doesn't reference works in progress.

I don't think the lifecycle stages of a docker image should depend on what language the code inside of it is written in, at least not within a single project, like Acumos.  But I don't see the piece for the stage-release step for non-maven builds, right now.

Patch-set: 1
Reviewer: Gerrit User 288 <288@655300a2-7a56-4504-9153-881edd907bea>