Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> do we need to up the version for lum-server for this change ?

I would say absolutely not.  This is not used or referenced by any of the versioned components of lum.  Not by the client, not by lum-db, and not by lum-server and therefore does not change the content of any jar or image that is part of lum-server.

Patch-set: 1
Reviewer: Gerrit User 288 <288@655300a2-7a56-4504-9153-881edd907bea>